Extracting functions to other files
