+++ src/pytest_devlife/constants.py
@@ -1,0 +1,10 @@
+SECTION_NAME = 'Developer Life'
+# FONT CONSTANTS
+BLUE = '\033[94m'
+CYAN = '\033[96m'
+GREEN = '\033[92m'
+WARNING = '\033[93m'
+FAIL = '\033[91m'
+ENDC = '\033[0m'
+BOLD = '\033[1m'
+UNDERLINE = '\033[4m'
